消除编译警告， 调整 Makefile 配置

+++ os_util.cpp
@@ -123,7 +123,7 @@ namespace nsp {
             return ::posix__rm(dir.c_str());
         }
 
-        int gettid() {
+        long gettid() {
             return ::posix__gettid();
         }
